+++ Encryptions/RC4.h
@@ -7,11 +7,11 @@
 
 class RC4 : public SymAlg{
     private:
-        uint8_t s_e[256], i_e, j_e,         // encryption SBox and "pointers"
-                s_d[256], i_d, j_d;         // decryption SBox and "pointers"
+        uint8_t s_e[256], i_e, j_e,      // encryption SBox and "pointers"
+                s_d[256], i_d, j_d;      // decryption SBox and "pointers"
 
-        void ksa(const std::string & k);    // key scheduling algorithm
-        int prga(const char mode);          // pseudo random generation algorithm
+        void ksa(const std::string & k); // key scheduling algorithm
+        uint8_t prga(const char mode);   // pseudo random generation algorithm
 
     public:
         RC4();
